pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Nikita Renjov <n.renjov@cft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.8.0"
cryptography = "^42.0.8"
fastapi = "^0.111.1"
bcrypt = "^4.1.3"
pydantic-settings = "^2.4.0"
brotli = "^1.1.0"
aiokafka = "^0.11.0"
aiofiles = "^24.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
asyncpg = "^0.29.0"
alembic = "^1.13.2"
prometheus-client = "^0.20.0"
jaeger-client = "^4.8.0"
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.19.2"
pytest = "^8.2.2"
black = "^24.4.2"
mypy = "^1.10.1"
coverage = "^7.5.4"
coverage-threshold = "^0.4.4"
pytest-asyncio = "^0.23.8"

[tool.isort]
multi_line_output = '3'
include_trailing_comma = 'True'
force_grid_wrap = '0'
use_parentheses = 'True'
ensure_newline_before_comments = 'True'
line_length = '79'

[tool.pytest.ini_options]
pythonpath = ". src"
testpaths = ["src/tests"]

[tool.mypy]
disable_error_code = 'import-untyped'

[tool.coverage.run]
command_line = "-m pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
